refactor: remove toolkit in place of static methods
